Ensure display of unsigned char variables.
We can do this as a static_assert checks if variables' type is integral.

+++ sdd/dd/node.hh
@@ -158,7 +158,7 @@ template <typename C, typename Valuation>
 std::ostream&
 operator<<(std::ostream& os, const node<C, Valuation>& n)
 {
-  os << n.variable() << "[";
+  os << static_cast<std::size_t>(n.variable()) << "[";
   std::for_each( n.begin(), n.end() - 1
                , [&](const arc<C, Valuation>& a)
                     {os << a.valuation() << " --> " << a.successor() << " || ";});
